requirements.txt
pytest-playwright
pylint
pytest-cov
black
